+++ pennylane_qiskit/qiskit_device.py
@@ -144,11 +144,12 @@ class QiskitDevice(QubitDevice, abc.ABC):
         "The analytic calculation of expectations, variances and "
         "probabilities is only supported on statevector backends, not on the {}. "
         "Such statistics obtained from this device are estimates based on samples."
+        "The simulation will be run with shots=1024."
     )
 
     _eigs = {}
 
-    def __init__(self, wires, provider, backend, shots=1024, **kwargs):
+    def __init__(self, wires, provider, backend, shots=None, **kwargs):
 
         super().__init__(wires=wires, shots=shots)